Support for large files (filesystem)

+++ src/fs/file.cpp
@@ -244,7 +244,7 @@ File &operator>>(File &fp, unsigned char *s)
 	return fp;
 }
 
-bool File::blockRead(void *buffer, unsigned int offset, size_t size)
+bool File::blockRead(void *buffer, uint64_t offset, uint64_t size)
 {
 	if (tell() != offset)
 	{
@@ -259,10 +259,10 @@ bool File::blockRead(void *buffer, unsigned int offset, size_t size)
 bool copyFile(File *const input, File *const output)
 {
 	input->rewind();
-	size_t toCopy = input->size();
-	const size_t bufferSize = 1 * 1024 * 1024;
+	uint64_t toCopy = input->size();
+	const uint64_t bufferSize = 1 * 1024 * 1024;
 	uint8_t *buffer = new uint8_t[bufferSize];
-	for (size_t readed = 0; (readed = input->read((char *)buffer, 1, std::min(bufferSize, toCopy))) != 0; toCopy -= readed)
+	for (uint64_t readed = 0; (readed = input->read((char *)buffer, 1, std::min(bufferSize, toCopy))) != 0; toCopy -= readed)
 	{
 		output->write(buffer, 1, readed);
 	}
